refactor: move some file and create folders
