+++ src/routes/graphql/root/mutation.ts
@@ -1,7 +1,7 @@
 import { GraphQLObjectType } from "graphql";
-import { userMutation } from "./user/user.js";
-import { PostMutation } from "./post/post.js";
-import { profileMutation } from "./profile/profile.js";
+import { userMutation } from "../mutation/user.js";
+import { PostMutation } from "../mutation/post.js";
+import { profileMutation } from "../mutation/profile.js";
 
 export const mutation = new GraphQLObjectType({
     name: 'Mutation',
