+++ src/routes/graphql/root/query.ts
@@ -1,8 +1,8 @@
 import { GraphQLObjectType } from "graphql";
-import { userQuery } from "./user.js";
-import { memberTypeQuery } from "./member.js";
-import { postQuery } from "./post.js";
-import { profileQuery } from "./profile.js";
+import { userQuery } from "../queries/user.js";
+import { memberTypeQuery } from "../queries/member.js";
+import { postQuery } from "../queries/post.js";
+import { profileQuery } from "../queries/profile.js";
 
 export const query = new GraphQLObjectType({
     name: 'Query',
